Remove excessive conversions from workitem to interface
